requirements.txt
alembic==1.13.1
bleach==6.1.0
Flask==3.0.3
SQLAlchemy==2.0.29
Werkzeug==3.0.2
WTForms==3.1.2
pillow~=11.0.0